spl_flags.h: Proper debug function declaration and implementation
Previously only the implementation was available through a separate
ifdef statement.

But now, user can define SPL_FLAGS_DEBUG in any .c file in addition to
the only .c file with SPL_FLAGS_IMPLEMENTATION.

NOTE: The user has to define SPL_FLAGS_DEBUG *also* in the file where
SPL_FLAGS_IMPLEMENTATION is defined in addition to the file where debug
function is to be used.